spec-tune: refactor logic for its own file
